requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
DateTime==5.1
idna==3.4
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.28.2
urllib3==1.26.15
zope.interface==6.0
